+++ src/poly1305/poly1305.h
@@ -12,6 +12,7 @@
 # pragma once
 
 #include <stddef.h>
+#include <stdint.h>
 
 #define POLY1305_BLOCK_SIZE  16
 #define POLY1305_DIGEST_SIZE 16
@@ -20,7 +21,7 @@
 typedef struct poly1305_context POLY1305;
 
 typedef void (*poly1305_blocks_f) (void *ctx, const uint8_t *inp,
-                                   size_t len, uint32_t padbit);
+                                   uint32_t len, uint32_t padbit);
 typedef void (*poly1305_emit_f) (void *ctx, uint8_t mac[16],
                                  const uint32_t nonce[4]);
 
@@ -31,7 +32,7 @@ struct poly1305_context {
                          * configurations */
     uint32_t nonce[4];
     uint8_t data[POLY1305_BLOCK_SIZE];
-    size_t num;
+    uint32_t num;
     struct {
         poly1305_blocks_f blocks;
         poly1305_emit_f emit;
@@ -40,7 +41,7 @@ struct poly1305_context {
 
 size_t Poly1305_ctx_size(void);
 void Poly1305_Init(POLY1305 *ctx, const uint8_t key[32]);
-void Poly1305_Update(POLY1305 *ctx, const uint8_t *inp, size_t len);
+void Poly1305_Update(POLY1305 *ctx, const uint8_t *inp, uint32_t len);
 void Poly1305_Final(POLY1305 *ctx, uint8_t mac[16]);
 
 #endif /* OSSL_CRYPTO_POLY1305_H */
